requirements.txt
# ---- Core API / utils
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3

# Flask deps compatible with Flask 3
click>=8.1.3,<9
Werkzeug>=3.0.0,<4
itsdangerous>=2.1.2,<3
Jinja2>=3.1.4,<4
MarkupSafe>=2.1.3,<3
blinker>=1.7,<2
urllib3>=2.2,<3
certifi>=2024.2.2
charset-normalizer>=3.3.2
idna>=3.6

# ---- Streamlit UI
streamlit==1.36.0
pandas==2.2.2

# ---- Gemini (chatbot)
google-generativeai==0.7.2
